protect multiple cleanup

+++ python/session.py
@@ -20,6 +20,7 @@ class Session(object):
     '''
     Some utilities helping to manage the session data
     '''
+    clean_in_progress = False
     
     def __init__(self, file):
         self.obsid=None
@@ -30,6 +31,7 @@ class Session(object):
         print(f"build session folder {self.path}")
         os.mkdir(self.path )
         self.filepath  = os.path.join(self.path, self.filename)
+        print(f"save {self.filename} in {self.filepath}")
         file.save(self.filepath)
         
         
@@ -106,21 +108,28 @@ class Session(object):
 
     @staticmethod
     def clean_up(delay_hours):
-        now = time.time()
-        old = now - (delay_hours*3600)
-        # remove older sessions
-        for root, dirs, _ in os.walk(PATHTO.sessions, topdown=False):
-            for _dir in dirs:
-                dir_path = os.path.join(root, _dir)
-                if os.path.getmtime(dir_path) < old:
-                    print(f"rm session {_dir}")
-                    shutil.rmtree(dir_path)     
-        
-        # remove older sessions
-        list_precomputed_obsids = os.listdir(PATHTO.precomputed_obsids)
-        if len(list_precomputed_obsids) >= 30:
-            oldest_file = min(list_precomputed_obsids, key=os.path.getctime)
-            print(f"rm cached file {oldest_file}")
-            os.remove(os.path.abspath(oldest_file))
-       
+        if Session.clean_in_progress:
+            return
+        Session.clean_in_progress = True
+        try:
+            now = time.time()
+            old = now - (delay_hours*3600)
+            # remove older sessions
+            for root, dirs, _ in os.walk(PATHTO.sessions, topdown=False):
+                for _dir in dirs:
+                    dir_path = os.path.join(root, _dir)
+                    if os.path.getmtime(dir_path) < old:
+                        print(f"rm session {_dir}")
+                        shutil.rmtree(dir_path)     
+            
+            # remove older sessions
+            list_precomputed_obsids = os.listdir(PATHTO.precomputed_obsids)
+            if len(list_precomputed_obsids) >= 30:
+                oldest_file = min(list_precomputed_obsids, key=os.path.getctime)
+                print(f"rm cached file {oldest_file}")
+                os.remove(os.path.abspath(oldest_file))
+        except Exception:
+            traceback.print_exc(file=sys.stdout)
+        Session.clean_in_progress = False
+
